NCSD-2166: Handle apis not existing

+++ PSScripts/Remove-ApisFromApiManagement.ps1
@@ -33,6 +33,9 @@ $Context = New-AzApiManagementContext -ResourceGroupName $ApimResourceGroup -Ser
 
 foreach ($ApiToRemove in $ApisToRemove) {
     $Api = Get-AzApiManagementApi -Context $Context -Name $ApiToRemove
-    Remove-AzApiManagementApi -Context $Context -ApiId $Api.ApiId
+
+    if($null -ne $Api) {
+        Remove-AzApiManagementApi -Context $Context -ApiId $Api.ApiId
+    }
 }
